+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+- [BREAKING] Remove `Langchain::Assistant#clear_thread!` method
 - [BREAKING] `Langchain::Messages::*` namespace had migrated to `Langchain::Assistant::Messages::*`
 - [BREAKING] Modify `Langchain::LLM::AwsBedrock` constructor to pass model options via default_options: {...}
 - Minor improvements to the Langchain::Assistant class